remote_llm/requirements.txt
accelerate==1.0.0
bitsandbytes==0.44.1
kserve==0.13.1
torch==2.4.1
transformers==4.45.2
sentencepiece
sentence_transformers
pyyaml
